+++ BE/exceed/src/main/java/com/gaebaljip/exceed/security/domain/JwtResolver.java
@@ -1,5 +1,6 @@
-package com.gaebaljip.exceed.security;
+package com.gaebaljip.exceed.security.domain;
 
+import com.gaebaljip.exceed.security.AuthConstants;
 import io.jsonwebtoken.*;
 import io.jsonwebtoken.io.Decoders;
 import io.jsonwebtoken.security.Keys;
